rename lvl so it is in order
