+++ blockly/src/level/produs/Chapter14Level.java
@@ -1,4 +1,4 @@
-package level.produs.chapter1;
+package level.produs;
 
 import contrib.components.LeverComponent;
 import core.Entity;
@@ -11,7 +11,7 @@ import entities.MiscFactory;
 import java.util.List;
 import level.BlocklyLevel;
 
-public class Chapter13Level extends BlocklyLevel {
+public class Chapter14Level extends BlocklyLevel {
 
   private DoorTile door;
   private LeverComponent switch1, switch2;
@@ -24,9 +24,9 @@ public class Chapter13Level extends BlocklyLevel {
    * @param designLabel The design label for the level.
    * @param customPoints The custom points of the level.
    */
-  public Chapter13Level(
+  public Chapter14Level(
       LevelElement[][] layout, DesignLabel designLabel, List<Coordinate> customPoints) {
-    super(layout, designLabel, customPoints, "Chapter13");
+    super(layout, designLabel, customPoints, "Kapitel 1: Level 4");
   }
 
   @Override
