+++ blockly/src/level/produs/Chapter15Level.java
@@ -1,4 +1,4 @@
-package level.produs.chapter1;
+package level.produs;
 
 import static level.Utils.cameraFocusOn;
 
@@ -12,7 +12,7 @@ import entities.MiscFactory;
 import java.util.List;
 import level.BlocklyLevel;
 
-public class Chapter14Level extends BlocklyLevel {
+public class Chapter15Level extends BlocklyLevel {
 
   /**
    * Call the parent constructor of a tile level with the given layout and design label. Set the
@@ -22,9 +22,9 @@ public class Chapter14Level extends BlocklyLevel {
    * @param designLabel The design label for the level.
    * @param customPoints The custom points of the level.
    */
-  public Chapter14Level(
+  public Chapter15Level(
       LevelElement[][] layout, DesignLabel designLabel, List<Coordinate> customPoints) {
-    super(layout, designLabel, customPoints, "Chapter14");
+    super(layout, designLabel, customPoints, "Kapitel 1: Level 5");
   }
 
   @Override
